backend/pyproject.toml
[project]
name = "modern-board-backend"
version = "0.1.0"
description = "Modern bulletin board application backend"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "wagtail>=7.2",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "gunicorn>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.9.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.4",
    "pyright>=1.1.390",
    "pre-commit>=4.0.0",
]

[tool.setuptools]
packages = ["api", "config"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

# Ruff configuration
[tool.ruff]
line-length = 88
target-version = "py313"
extend-exclude = ["migrations", "static", "media", ".venv"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["api", "config"]

# pytest configuration
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["api/tests"]
addopts = [
    "--strict-markers",
    "--verbose",
    "--tb=short",
]

# Coverage configuration
[tool.coverage.run]
source = ["api"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__pycache__/*",
    "*/static/*",
    "*/media/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

# Pyright configuration
[tool.pyright]
include = ["api", "config"]
exclude = [
    "**/migrations",
    "**/__pycache__",
    "**/tests",
    ".venv",
]
pythonVersion = "3.13"
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
